LayoutTests/platform/blackberry/fast/forms/listbox-selection-expected.txt

<select> selection test for mouse events and keyevents.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


1) Select one item with mouse (no previous selection)
FAIL selectionPattern("sl1") should be 10000. Was 00000.
2) Select one item with mouse (with previous selection)
FAIL selectionPattern("sl2") should be 10000. Was 01000.
3) Select one item with the keyboard (no previous selection)
PASS selectionPattern("sl3") is "00001"
4) Select one item with the keyboard (with previous selection)
PASS selectionPattern("sl4") is "00100"
5) Attempt to select an item cmd-clicking
FAIL selectionPattern("sl5") should be 01000. Was 00100.
6) Attempt to select a range shift-clicking
FAIL selectionPattern("sl6") should be 01000. Was 00010.
7) Attempt to select a range with the keyboard
PASS selectionPattern("sl7") is "00100"
8) Select one item with mouse (no previous selection)
FAIL selectionPattern("sl8") should be 10000. Was 00000.
9) Select one item with mouse (with previous selection)
FAIL selectionPattern("sl9") should be 10000. Was 01000.
10) Select one item with the keyboard (no previous selection)
PASS selectionPattern("sl10") is "00001"
11) Select one item with the keyboard (with previous selection)
PASS selectionPattern("sl11") is "00100"
12) Select an item cmd-clicking
FAIL selectionPattern("sl12") should be 01100. Was 00100.
13) Select a range shift-clicking
FAIL selectionPattern("sl13") should be 11110. Was 10000.
14) Select a range with the keyboard
FAIL selectionPattern("sl14") should be 01100. Was 00100.
PASS successfullyParsed is true

TEST COMPLETE
